+++ cosec-cocache/src/main/kotlin/me/ahoo/cosec/cache/RolePermissionCache.kt
@@ -11,7 +11,7 @@
  * limitations under the License.
  */
 
-package me.ahoo.cosec.redis
+package me.ahoo.cosec.cache
 
 import me.ahoo.cache.api.Cache
 import me.ahoo.cosec.Delegated
