+++ settings.gradle.kts
@@ -18,7 +18,7 @@ include(":cosec-dependencies")
 include(":cosec-api")
 include(":cosec-core")
 include(":cosec-jwt")
-include(":cosec-redis")
+include(":cosec-cocache")
 include(":cosec-social")
 include(":cosec-webmvc")
 include(":cosec-webflux")
